Fixes
git-svn-id: svn+ssh://svn.cern.ch/reps/dirac/LHCbDIRAC/trunk@73443 4525493e-7705-40b1-a816-d608a930855b

+++ LHCbDIRAC/Workflow/Modules/ModulesUtilities.py
@@ -80,7 +80,7 @@ def getEventsToProduce( CPUe, CPUTime = None, CPUNormalizationFactor = None,
   gLogger.verbose( "Without limits, we can produce %d events" % eventsToProduce )
 
   gLogger.info( "We can produce %d events" % eventsToProduce )
-  willProduce = eventsToProduce * 0.8
+  willProduce = int( eventsToProduce * 0.8 )
   gLogger.info( "But we take a conservative approach, so 80%% of those: %d" % willProduce )
 
   if maxNumberOfEvents:
